+++ attachment.php
@@ -22,10 +22,13 @@
                     <span class="image-date">Uploaded on <time datetime="<?=get_the_date('c');?>" itemprop="datePublished"><?php the_date(); ?></time></span>
                     <span class="image-author">by <a href="<?=get_author_posts_url( get_the_author_meta( 'ID' ), get_the_author_meta( 'user_nicename' ) );?>" itemprop="author" rel="author"><?php the_author(); ?></a></span>
                     <br>
+                    <span class="image-credit">Credit: <b><?=get_credit_text();?></b></span>
+                    <br>
                     <span class="image-category">Category: <b><?php the_category(' '); ?></b></span>
                     <br>
                     <span class="image-filesize" itemprop="size"><?=image_metadata(wp_get_attachment_url(get_the_ID()));?></span>
                 </p>
+
                 <div class="image-share"><?=blog_post_share(); ?></div>
                 <div class="image-links"><i>Related Post</i><br> <a href="<?=get_permalink( get_post_parent(get_the_ID()) );?>"><?php if (has_post_parent(get_the_ID())) echo get_the_title( get_post_parent(get_the_ID()) ); else echo "Not attached to any posts"; ?></a></div>
             </div>